Restructuring and adding tests.
