+++ graylog2-server/src/test/java/org/graylog2/indexer/IndexMappingFactoryTest.java
@@ -17,6 +17,7 @@
 package org.graylog2.indexer;
 
 import com.github.zafarkhaja.semver.ParseException;
+import com.github.zafarkhaja.semver.Version;
 import com.google.gson.JsonObject;
 import io.searchbox.client.JestClient;
 import io.searchbox.client.JestResult;
@@ -46,42 +47,11 @@ public class IndexMappingFactoryTest {
 
     @Mock
     private JestClient jestClient;
-    private IndexMappingFactory indexMappingFactory;
-
-    @Before
-    public void setUp() throws Exception {
-        indexMappingFactory = new IndexMappingFactory(jestClient);
-    }
-
-    @Test
-    public void createIndexMappingFailsIfElasticsearchIsUnavailable() throws Exception {
-        when(jestClient.execute(any(Ping.class))).thenThrow(IOException.class);
-
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
-                .isInstanceOf(ElasticsearchException.class)
-                .hasCauseInstanceOf(IOException.class);
-    }
-
-    @Test
-    public void createIndexMappingFailsIfElasticsearchResponseFailed() throws Exception {
-        final JestResult failedResult = mock(JestResult.class);
-        when(failedResult.isSucceeded()).thenReturn(false);
-        when(failedResult.getJsonObject()).thenReturn(new JsonObject());
-        when(jestClient.execute(any(Ping.class))).thenReturn(failedResult);
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
-                .isInstanceOf(ElasticsearchException.class)
-                .hasMessageStartingWith("Unable to retrieve Elasticsearch version")
-                .hasNoCause();
-    }
-
     @Test
     public void createIndexMappingFailsIfElasticsearchVersionIsTooLow() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("1.7.3"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
+        final IndexMappingFactory indexMappingFactory = new IndexMappingFactory(Version.valueOf("1.7.3"));
 
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
+        assertThatThrownBy(indexMappingFactory::createIndexMapping)
                 .isInstanceOf(ElasticsearchException.class)
                 .hasMessageStartingWith("Unsupported Elasticsearch version: 1.7.3")
                 .hasNoCause();
@@ -89,12 +59,9 @@ public class IndexMappingFactoryTest {
 
     @Test
     public void createIndexMappingFailsIfElasticsearch2VersionIsTooLow() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("2.0.0"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
+        final IndexMappingFactory indexMappingFactory = new IndexMappingFactory(Version.valueOf("2.0.0"));
 
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
+        assertThatThrownBy(indexMappingFactory::createIndexMapping)
                 .isInstanceOf(ElasticsearchException.class)
                 .hasMessageStartingWith("Unsupported Elasticsearch version: 2.0.0")
                 .hasNoCause();
@@ -102,36 +69,12 @@ public class IndexMappingFactoryTest {
 
     @Test
     public void createIndexMappingFailsIfElasticsearchVersionIsTooHigh() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("6.0.0"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
+        final IndexMappingFactory indexMappingFactory = new IndexMappingFactory(Version.valueOf("6.0.0"));
 
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
+        assertThatThrownBy(indexMappingFactory::createIndexMapping)
                 .isInstanceOf(ElasticsearchException.class)
                 .hasMessageStartingWith("Unsupported Elasticsearch version: 6.0.0")
                 .hasNoCause();
-    }
-
-    @Test
-    public void createIndexMappingFailsIfElasticsearchVersionIsInvalid() throws Exception {
-        final JestResult jestResult = mock(JestResult.class);
-        when(jestResult.isSucceeded()).thenReturn(true);
-        when(jestResult.getJsonObject()).thenReturn(buildVersionJsonObject("Foobar"));
-        when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
-
-        assertThatThrownBy(() -> indexMappingFactory.createIndexMapping())
-                .isInstanceOf(ElasticsearchException.class)
-                .hasMessageStartingWith("Unable to parse Elasticsearch version: Foobar")
-                .hasCauseInstanceOf(ParseException.class);
-    }
-
-    private static JsonObject buildVersionJsonObject(String foobar) {
-        final JsonObject versionObject = new JsonObject();
-        versionObject.addProperty("number", foobar);
-        final JsonObject jsonObject = new JsonObject();
-        jsonObject.add("version", versionObject);
-        return jsonObject;
     }
 
     @RunWith(Parameterized.class)
@@ -158,8 +101,6 @@ public class IndexMappingFactoryTest {
         private final String version;
         private final Class<? extends IndexMapping> expectedMapping;
 
-        @Mock
-        private JestClient jestClient;
         private IndexMappingFactory indexMappingFactory;
 
 
@@ -170,17 +111,11 @@ public class IndexMappingFactoryTest {
 
         @Before
         public void setUp() throws Exception {
-            indexMappingFactory = new IndexMappingFactory(jestClient);
+            indexMappingFactory = new IndexMappingFactory(Version.valueOf(this.version));
         }
 
         @Test
         public void test() throws Exception {
-            final JestResult jestResult = mock(JestResult.class);
-            when(jestResult.isSucceeded()).thenReturn(true);
-            final JsonObject jsonObject = buildVersionJsonObject(version);
-            when(jestResult.getJsonObject()).thenReturn(jsonObject);
-            when(jestClient.execute(any(Ping.class))).thenReturn(jestResult);
-
             assertThat(indexMappingFactory.createIndexMapping()).isInstanceOf(expectedMapping);
         }
     }
