+++ graylog2-server/src/test/java/org/graylog2/indexer/indices/IndicesGetAllMessageFieldsTest.java
@@ -28,6 +28,7 @@ import org.elasticsearch.action.admin.indices.get.GetIndexRequest;
 import org.elasticsearch.action.admin.indices.get.GetIndexResponse;
 import org.graylog2.AbstractESTest;
 import org.graylog2.audit.NullAuditEventSender;
+import org.graylog2.bindings.providers.ElasticsearchVersionProvider;
 import org.graylog2.indexer.IndexMappingFactory;
 import org.graylog2.indexer.messages.Messages;
 import org.graylog2.plugin.system.NodeId;
@@ -53,9 +54,10 @@ public class IndicesGetAllMessageFieldsTest extends AbstractESTest {
     @Before
     public void setUp() throws Exception {
         super.setUp();
+        final ElasticsearchVersionProvider elasticsearchVersionProvider = new ElasticsearchVersionProvider(jestClient());
         indices = new Indices(jestClient(),
                 new Gson(),
-                new IndexMappingFactory(jestClient()),
+                new IndexMappingFactory(elasticsearchVersionProvider.get()),
                 new Messages(new MetricRegistry(), jestClient()),
                 mock(NodeId.class),
                 new NullAuditEventSender(),
