+++ graylog2-server/src/test/java/org/graylog2/indexer/indices/IndicesTest.java
@@ -46,6 +46,7 @@ import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.elasticsearch.common.compress.CompressedXContent;
 import org.graylog2.AbstractESTest;
 import org.graylog2.audit.NullAuditEventSender;
+import org.graylog2.bindings.providers.ElasticsearchVersionProvider;
 import org.graylog2.indexer.IndexMapping;
 import org.graylog2.indexer.IndexMappingFactory;
 import org.graylog2.indexer.IndexNotFoundException;
@@ -118,7 +119,8 @@ public class IndicesTest extends AbstractESTest {
     public void setUp() throws Exception {
         super.setUp();
         eventBus = new EventBus("indices-test");
-        indexMappingFactory = new IndexMappingFactory(jestClient());
+        final ElasticsearchVersionProvider elasticsearchVersionProvider = new ElasticsearchVersionProvider(jestClient());
+        indexMappingFactory = new IndexMappingFactory(elasticsearchVersionProvider.get());
         indices = new Indices(jestClient(),
                 new Gson(),
                 indexMappingFactory,
